Fixed ssize_t type definition for mpg123 (affects MSVC x64 only)
According to MinGW headers (which is used for compiling libmpg123) ssize_t type must be 64-bit in size on 64-bit Windows
Moreover it was impossible to build GZDoom with the recent versions of libmpg123 because of wrong type redifinition
Tested MP3 music on Deus Vult II with Release x64 and OpenAL backend

+++ src/sound/mpg123_decoder.h
@@ -6,7 +6,8 @@
 #ifdef HAVE_MPG123
 
 #ifdef _MSC_VER
-typedef int ssize_t;
+#include <stddef.h>
+typedef ptrdiff_t ssize_t;
 #endif
 #include "mpg123.h"
